refactor: move persistence-related classes to correct package
