+++ src/main/java/io/retel/ariproxy/persistence/PersistenceStore.java
@@ -1,7 +1,12 @@
 package io.retel.ariproxy.persistence;
 
+import akka.actor.typed.ActorRef;
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
+import io.retel.ariproxy.metrics.MetricsServiceMessage;
 import io.vavr.concurrent.Future;
 import io.vavr.control.Option;
+import io.vavr.control.Try;
 
 public interface PersistenceStore {
